requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
autoviz==0.1.905
babel==2.17.0
beautifulsoup4==4.13.4
bitarray==3.4.1
bleach==6.2.0
bokeh==3.4.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
ckzg==2.1.1
click==8.1.8
cloudpickle==3.1.1
colorcet==3.1.0
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
cytoolz==1.0.1
dacite==1.9.2
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
emoji==2.14.1
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
exceptiongroup==1.3.0
executing==2.2.0
Faker==37.1.0
fastjsonschema==2.21.1
featuretools==1.31.0
fonttools==4.58.0
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2025.3.2
h11==0.16.0
hexbytes==1.3.1
holidays==0.72
holoviews==1.20.2
htmlmin==0.1.12
httpcore==1.0.9
httpx==0.28.1
hvplot==0.11.3
idna==3.10
ImageHash==4.3.1
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.18.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.0
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.7
linkify-it-py==2.0.3
llvmlite==0.43.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
mistune==3.1.3
multidict==6.4.4
multimethod==1.12
narwhals==1.39.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
notebook_shim==0.2.4
numba==0.60.0
numpy==2.0.2
overrides==7.7.0
packaging==23.2
pandas==2.2.3
pandas-dq==1.29
pandas-profiling==3.6.6
pandocfilters==1.5.1
panel==1.4.5
param==2.2.0
parsimonious==0.10.0
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
phik==0.12.4
pillow==11.2.1
platformdirs==4.3.8
plotly==6.0.1
prometheus_client==0.21.1
prompt_toolkit==3.0.51
propcache==0.3.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.29
py-solc-x==2.0.3
pyamg==5.2.1
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
pytz==2025.2
pyunormalize==16.0.0
pyviz_comms==3.0.4
PyWavelets==1.6.0
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rlp==4.1.0
rpds-py==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
statsmodels==0.14.4
sweetviz==2.3.1
terminado==0.18.1
textblob==0.19.0
threadpoolctl==3.6.0
tinycss2==1.4.0
tomli==2.2.1
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typeguard==4.4.2
types-python-dateutil==2.9.0.20241206
types-requests==2.32.0.20250515
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uc-micro-py==1.0.3
uri-template==1.3.0
urllib3==2.4.0
visions==0.8.1
wcwidth==0.2.13
web3==7.11.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
woodwork==0.31.0
wordcloud==1.9.4
xgboost==1.6.2
xlrd==2.0.1
xyzservices==2025.4.0
yarl==1.20.0
ydata-profiling==4.16.1
zipp==3.21.0
